@@ -1,10 +1,19 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
+export default {
   content: [
-    "./src/**/*.{js,jsx,ts,tsx}",
+    "./index.html",
+    "./src/**/*.{js,ts,jsx,tsx}",
   ],
   theme: {
-    extend: {},
+    extend: {
+      fontFamily:{
+        "Montserrat": ["Montserrat", "sans-serif"]
+      },
+      colors:{
+        headerColor: "rgba(60, 15, 15, 1)",
+        headerTextColor: "rgba(227, 237, 107, 1)"
+      }
+    },
   },
   plugins: [],
 }
